pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omicsguard"
version = "0.1.0"
description = "A Serverless-Ready Metadata Validator for GA4GH Genomic Standards"
readme = "README.md"
authors = [
  { name="Anil Bodepudi", email="anilbodepudi@example.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.9"
dependencies = [
    "jsonschema>=4.0.0",
    "click>=8.0.0",
    "requests>=2.25.0",
    "pyyaml>=6.0",
]

[tool.setuptools.package-data]
omicsguard = ["schemas/*.json"]

[project.scripts]
omicsguard = "omicsguard.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "coverage",
]
